[refactor] users informaiton form as component
